+++ src/app/constants/index.ts
@@ -1,7 +1,5 @@
 import { DbClient } from './DbClient'
 import { ServicePorts } from './ports'
-import { ActionNames } from './names/actions'
-import { ModuleNames } from './names/modules'
 import { AuthSettingKeys } from './setting-keys/auth'
 import { CacheSettingKeys } from './setting-keys/cache'
 import { DbSettingKeys } from './setting-keys/database'
@@ -13,8 +11,6 @@ import { WebSettingKeys } from './setting-keys/web'
 export type Constants = {
     DbClient: typeof DbClient,
     ServicePorts: typeof ServicePorts,
-    ActionNames: typeof ActionNames,
-    ModuleNames: typeof ModuleNames,
     AuthSettingKeys: typeof AuthSettingKeys,
     CacheSettingKeys: typeof CacheSettingKeys,
     DbSettingKeys: typeof DbSettingKeys,
@@ -25,10 +21,8 @@ export type Constants = {
 }
 
 export const constants: Constants = {
-    ActionNames,
     DbClient,
     ServicePorts,
-    ModuleNames,
     AuthSettingKeys,
     CacheSettingKeys,
     DbSettingKeys,