+++ src/app/models/id/SingleId.ts
@@ -1,0 +1,18 @@
+import { IdBase } from './IdBase'
+
+
+export class SingleId extends IdBase {
+
+    constructor(
+        public id: string
+    ) {
+        super()
+    }
+
+    /**
+     * @override
+     */
+    public toArray(): string[] {
+        return [ this.id ]
+    }
+}